chore: testing the release of 2.0 to go back to latest@main releases
After merging next into main we were having trouble releasing in the latest dist-tag from npm
because of the way we used the next dist-tag. It's meant for breaking change releases but we were
using it for minor/patch. Let's try to get a clean state for 2.0

BREAKING CHANGE: Starting fresh